restructure run_stats
in order to simplify the ability to add statistics to
new command, the run_stats class were restructured
